backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6

# LangChain and AI
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langgraph>=0.0.50
openai>=1.0.0

# Document Processing
PyPDF2>=3.0.0
pytesseract>=0.3.10
python-docx>=1.0.0
Pillow>=10.0.0

# Medical Imaging
pydicom>=2.4.0
opencv-python>=4.8.0

# ML/AI
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Utilities
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi>=0.1.9
python-magic>=0.4.27

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
